Move data files to own directory
Also, don't use .jsx for non-React files.
