+++ uclapi-frontend/data/homepage_constants.jsx
@@ -1,6 +1,3 @@
-// React
-
-
 export const endpoints = [
   {
     name: `/oauth`,
