+++ uclapi-frontend/pages/dashboard.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/jsx-no-bind */
-import { styles as dashboardStyles } from '@/components/layout/data/dashboard_styles.jsx'
+import { styles as dashboardStyles } from '@/data/dashboard_styles'
 import {
   CardView, Column, ConfirmBox,
   Container, Row,
