+++ uclapi-frontend/pages/marketplace/index.jsx
@@ -6,7 +6,7 @@ import "react-responsive-carousel/lib/styles/carousel.min.css"
 import styles from "@/styles/Marketplace.module.scss"
 
 // Grab titles and descriptions of app
-import { allApps } from '@/components/layout/data/app_pages.jsx'
+import { allApps } from '@/data/app_pages.jsx'
 // Common Components
 import {
   CardView, Container,
